ui stage 2

+++ bangbangboom/Controllers/UserController.cs
@@ -11,6 +11,7 @@ using Microsoft.Net.Http.Headers;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.IO;
 using System.Linq;
 using System.Net.Mime;
 using System.Threading.Tasks;
@@ -81,9 +82,16 @@ namespace bangbangboom.Controllers
             var hash = user?.ProfileFileHash;
             if (hash == null)
                 return StatusCode(404);
-            var fs = fileProvider.GetFileByHash(hash);
-            return File(fs, "image/jpeg", null,
-                EntityTagHeaderValue.Parse(new StringSegment('"' + hash + '"')), true);
+            try
+            {
+                var fs = fileProvider.GetFileByHash(hash);
+                return File(fs, "image/jpeg", null,
+                    EntityTagHeaderValue.Parse(new StringSegment('"' + hash + '"')), true);
+            }
+            catch (FileNotFoundException)
+            {
+                return StatusCode(404);
+            }
         }
 
         [Authorize]
@@ -99,7 +107,8 @@ namespace bangbangboom.Controllers
                 return StatusCode(400);
             var user = await userManager.GetUserAsync(User);
             var hash = await fileProvider.SaveFileAsync(jpg);
-            fileProvider.DeleteFile(user.ProfileFileHash);
+            if (user.ProfileFileHash != null)
+                fileProvider.DeleteFile(user.ProfileFileHash);
             user.ProfileFileHash = hash;
             await context.SaveChangesAsync();
             return Ok();
